fixed file structure/ UI
